requirements.txt
python-telegram-bot==20.7
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.32.3
ruff==0.6.8
mypy==1.10.0
